千葉さんにご教示いただいた内容の修正
